Add Comment.nvim integration code into this plugin
This should make it simpler for users to set up the integration, as well
as allow us to add more features to it in the future. Users wouldn't
need to maintain the integration themselves.

This does mean that if there are any breaking changes inside
Comment.nvim, we would need to release a fix for the integration fast.