fix(jsx): correctly comment first line of JSX
The first and last lines of JSX should be commented with `// %s`, rather
than `{/* %s */}`. However, the current way of configuring the plugin
does not allow us to check if the current node is a `jsx_element` with a
parent of `parenthesized_expression`. We could allow checking the parent
node in the configuration, but that would be a bit tricky and a lot of
work.

Since this is (hopefully) a 1-time exception, then it seems okay to add
an explicit check for this. If we have similar cases in the future
though, then we should think of making this configurable somehow.

Fixes #29