feat: support nvim-treesitter 1.0
`nvim-treesitter` modules are removed in 1.0 and Treesitter plugins need
to set themselves up on their own. This commit adds a `setup` function
which sets up autocommands to initialize the plugins in buffers with
Treesitter enabled.

The second part is using the new Treesitter functions from Neovim if
they are available, falling back to `nvim-treesitter` functions. Both
`nvim-treesitter` < 1.0 and >= 1.0 should be supported now.

Related to #72