fix: do not fail if Treesitter Node name matches string method or field
Example problem case with the Haskell Treesitter parser:

1. There is a Treesitter Node called "match"
2. The config for Haskell is a string: `haskell = '-- %s',`
3. In our code, we do `language_config[<ts-node-name>]`, but if
   `language_config` is a `string`, then accessing its `match` key would
   return the `string.match` function:
   ```lua
   local my_variable = 'hello'
   vim.print(my_variable.match)
   -- Outputs a function
   ```
4. We then try to access a field of the function, which causes a runtime
   error :(

To fix this, we can avoid a lot of the logic in `check_node` by just
checking if the `language_config` is a `string` and returning early in
that case. This should also bring a small performance boost because we
don't need to traverse so many Treesitter nodes in a lot of cases.

Related to #97 and #103